fix: seprete types from logics
