+++ packages/components/form/src/components/actions/submit/submit.types.ts
@@ -1,7 +1,7 @@
 import { FieldValues } from 'react-hook-form';
-import { Form } from '../../../hooks/useForms/useForms';
 import { FormId } from '../../../types/public.types';
 import { ActionProps } from '../action/action.types';
+import { Form } from '../../../hooks/useForms/useForms.types';
 
 export type SubmitFieldProps = ActionProps & {
   formId: FormId
