+++ packages/components/form/src/hooks/usePropsController/usePropsController.types.ts
@@ -1,0 +1,6 @@
+export type SetProps = (id: string, props: any) => void;
+
+export type UseControllerState = {
+  propsController: Record<string, any>;
+  setProps: (id: string, props: any) => void;
+};
