+++ packages/components/form/src/hooks/useScript/useScript.types.ts
@@ -1,0 +1,9 @@
+import { FormId, Forms, Script } from '../../types/public.types';
+import { Form } from "../useForms/useForms.types";
+
+export type UseScriptProps = {
+  script?: Script;
+  formMethod: Form;
+  forms: Forms;
+  formId: FormId;
+};
