+++ packages/components/form/src/utils/formWrapper/formWrapper.types.tsx
@@ -1,0 +1,3 @@
+import { ReactNode } from "react";
+
+export type FormWrapperProps = { children: ReactNode; formId: string }
